pyproject.toml
[tool.poetry]
name = "opendata-bridge-ai-api"
version = "1.0.0"
description = "OpenData Bridge AI API"
authors = ["FooQoo <horseshoes.buzz.17@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "opendata_bridge_data_norm_api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
openai = "^1.11.1"
jupyter = "^1.0.0"
pillow = "^10.2.0"
python-multipart = "^0.0.7"
pydantic-settings = "^2.1.0"
pandas = "^2.2.0"
fastapi-utilities = "^0.2.0"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
black = "^24.0.0"
flake8 = "^7.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
